Give up (for now) on storing UTC timestamps in the database.
Storing UTC would be the right thing to do, but Python/Django make it difficult
(e.g. the return value of datetime.datetime.utcnow() is the UTC time as a naive
datetime, and when you insert it into Postgres, Postgres assumes server
timezone, so you've just screwed it up - you have to explicitly give every
datetime the UTC tzinfo). And since we want to display timestamps in server
timezone anyway, we're not trying to convert to localized per-user timezones,
currently this is just more trouble than it's worth. So back to server-timezone
everywhere, at least for now.